% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-rast.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,RasterArray-method}
\title{Aggregate raster cells in a \code{\link[via:RasterArray-class]{RasterArray}} object}
\usage{
\S4method{aggregate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[via:RasterArray-class]{RasterArray}}-class object.}

\item{...}{arguments passed to the \code{\link[terra]{aggregate}} function.}
}
\value{
An aggregated \code{\link[via:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the \code{\link[terra:rast]{SpatRaster}} class.
}
\examples{
library(terra)
ex <- rastex()
agg <- aggregate(ex, 30)
}
