% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.XArray}
\alias{as.XArray}
\alias{as.XArray,SfcArray-method}
\alias{as.XArray,SfArray-method}
\alias{as.SfcArray}
\alias{as.SfcArray,XArray-method}
\alias{as.SfcArray,SfArray-method}
\alias{as.SfArray}
\alias{as.SfArray,XArray-method}
\alias{as.SfArray,SfcArray-method}
\alias{as}
\title{Coerce into an \code{\link[via:SfcArray-class]{SfcArray}} or \code{\link[via:SfArray-class]{SfArray}} object}
\usage{
as.XArray(from)

\S4method{as.XArray}{SfcArray}(from)

\S4method{as.XArray}{SfArray}(from)

as.SfcArray(from)

\S4method{as.SfcArray}{XArray}(from)

\S4method{as.SfcArray}{SfArray}(from)

as.SfArray(from)

\S4method{as.SfArray}{XArray}(from)

\S4method{as.SfArray}{SfcArray}(from)
}
\arguments{
\item{from}{Either a \code{\link[via:SfcArray-class]{SfcArray}}, \code{\link[via:SfArray-class]{SfArray}} or \code{\link[via:XArray-class]{XArray}}-class object}
}
\value{
Either a \code{\link[via:SfcArray-class]{SfcArray}}, \code{\link[via:SfcArray-class]{SfArray}} or \code{\link[via:XArray-class]{XArray}}-class object
}
\description{
Coerce into an \code{\link[via:SfcArray-class]{SfcArray}} or \code{\link[via:SfArray-class]{SfArray}} object

Coerce into an SfArray or SfcArray object
}
