% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{colnames,VirtualArray-method}
\alias{colnames,VirtualArray-method}
\alias{colnames<-,VirtualArray-method}
\title{Column names of two-dimensional \code{\link{RasterArray}} or \code{\link{SpatialArray}}}
\usage{
\S4method{colnames}{VirtualArray}(x)

\S4method{colnames}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link{RasterArray}} or \code{\link{SpatialArray}} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of column names or \code{NULL}.
}
\description{
Get or set the column names of two-dimensional code{\link{RasterArray}} or \code{\link{SpatialArray}} objects
}
\examples{
data(coasts)
colnames(coasts)
colnames(coasts) <- c("a", "b")
}
