% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-combine.R
\name{c,VirtualArray-method}
\alias{c,VirtualArray-method}
\title{Combine a one-dimensional VirtualArray with other obejcts}
\usage{
\S4method{c}{VirtualArray}(x, ...)
}
\arguments{
\item{x}{\code{VirtualArray} object to combine wit other objects.}

\item{...}{additional objects to combine.}
}
\value{
A \code{\link[via:XArray-class]{VirtualArray}}-class object.
}
\description{
Methods sequences that start with an NA do not yet work.
}
