% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-xattrib.R
\name{dimlayer}
\alias{dimlayer}
\alias{dimlayer,RasterArray-method}
\title{Dimensions of layers in a VirtualArray object}
\usage{
dimlayer(x, ...)

\S4method{dimlayer}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{VirtualArray} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{numeric} vector with the number of rows and columns in the \code{VirtualArray}s.
}
\description{
The funcion will return the dimensions RasterLayers
}
