% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{dimnames,VirtualArray-method}
\alias{dimnames,VirtualArray-method}
\alias{dimnames<-,VirtualArray-method}
\title{Names of multidimensional \code{\link{RasterArray}} or \code{\link{SpatialArray}} objects.}
\usage{
\S4method{dimnames}{VirtualArray}(x)

\S4method{dimnames}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link{RasterArray}} or \code{\link{SpatialArray}} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{list} of \code{character} vectors or \code{NULL}.
}
\description{
Get or set the dimnames of multidimensional \code{\link{RasterArray}} or \code{\link{SpatialArray}} objects
}
\examples{
data(dems)
dimnames(dems)
data(coasts)
dimnames(coasts)
dimnames(coasts)[[2]] <- c("first", "second")
}
