% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-rast.R
\name{disagg}
\alias{disagg}
\alias{disagg,RasterArray-method}
\title{Disaggregate raster cells in a '\code{\link[via:RasterArray-class]{RasterArray}}'-class object}
\usage{
\S4method{disagg}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[via:RasterArray-class]{RasterArray}} class object.}

\item{...}{arguments passed to the \code{\link[terra]{disagg}} function.}
}
\value{
A disaggregated \code{\link[via:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the '\code{\link[terra:rast]{SpatRaster}}' class.
}
\examples{
ex <- rastex() 
disagg <- disagg(ex, 3)
}
