% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-replace.R
\name{[[<-,VirtualArray,ANY-method}
\alias{[[<-,VirtualArray,ANY-method}
\alias{[[<-,VirtualArray-method}
\alias{[[<-,VirtualArray,ANY,ANY-method}
\title{Replace elements of '\code{\link[via:XArray-class]{VirtualArray}}'-class objects.}
\usage{
\S4method{[[}{VirtualArray,ANY}(x, i) <- value
}
\arguments{
\item{x}{Object from a class derived from \code{\link[via:XArray-class]{VirtualArray}}.}

\item{i}{subscript of layers to replace.}

\item{value}{\code{character} vector.}
}
\value{
The function has no return value.
}
\description{
Double bracket \code{'[['} refers to layers' name in the names of the \code{@stack} member of the '\code{\link[via:XArray-class]{VirtualArray}}'. Use single brackets to replace elements based on their position in the '\code{\link[via:XArray-class]{VirtualArray}}'-class object.
}
