% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{oneRast}
\alias{oneRast}
\alias{twoRast}
\title{Example structures to demonstrate the capabilities of the package}
\usage{
oneRast()

twoRast()
}
\description{
Example structures to demonstrate the capabilities of the package
}
