% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-rast.R
\name{ext}
\alias{ext}
\alias{ext,RasterArray-method}
\title{Extent of a \code{\link[via:RasterArray-class]{RasterArray}} object}
\usage{
\S4method{ext}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[via:RasterArray-class]{RasterArray}}-class object.}

\item{...}{arguments passed to the \code{\link[terra]{ext}} function.}
}
\value{
An aggregated \code{\link[via:RasterArray-class]{RasterArray}}-class object.
}
\description{
The method is inherited from the \code{\link[terra:rast]{SpatRaster}} class.
}
\examples{
ex <- rastex()
extent <- ext(ex)
}
