% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{genarray}
\alias{genarray}
\title{Generalized ArrayRray}
\description{
Cool package to be written.
}
\details{
This is still the Beta version. As is R, this is free software and comes with ABSOLUTELY NO WARRANTY. Nevertheless, notes about found bugs and suggestions are more than welcome.
}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com) and Nussaibah B. Raja
}
