% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RastArray-base.R
\name{is.na.RastArray}
\alias{is.na.RastArray}
\title{Positions of missing values in a RasterArray object}
\usage{
\method{is.na}{RastArray}(x)
}
\arguments{
\item{x}{A \code{RasterArray} class object.}
}
\value{
A \code{logical} \code{vector}, \code{matrix} or \code{array} matching the structure of the \code{RastArray}.
}
\description{
The function behaves similar to the regular \code{is.na()} function applied to the proxy object of a \code{RastArray}.
}
\examples{
data(dems)
dems[2] <- NA
is.na(dems)

}
