% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/VirtualArray-xattrib.R
\name{layers}
\alias{layers}
\alias{layers,VirtualArray-method}
\title{Names of layers in the \code{stack} of a '\code{\link[via:XArray-class]{VirtualArray}}'-class object}
\usage{
layers(x, ...)

\S4method{layers}{VirtualArray}(x)
}
\arguments{
\item{x}{A \code{\link[via:XArray-class]{VirtualArray}}-derived class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{character} vector of names.
}
\description{
Names of layers in the \code{stack} of a '\code{\link[via:XArray-class]{VirtualArray}}'-class object
}
\examples{
# names of layers in the stack
data(exemplar)
layers(exemplar)
}
