% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{names,VirtualArray-method}
\alias{names,VirtualArray-method}
\alias{names<-,VirtualArray-method}
\title{Names of one-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class objects.}
\usage{
\S4method{names}{VirtualArray}(x)

\S4method{names}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link[via:XArray-class]{VirtualArray}}-derived class object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of names or \code{NULL}.
}
\description{
Get or set the names of one-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class objects
}
\examples{
ex <- rastex()
names(ex)
names(ex)[4] <- "weirdo"
# NULL
}
