% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nums}
\alias{nums}
\alias{colnums}
\alias{rownums}
\title{Names as numerics}
\usage{
nums(x)

colnums(x)

rownums(x)
}
\arguments{
\item{x}{Object with names, colnames or rownames attributes.}
}
\value{
Numeric vector.
}
\description{
The set of functions return names of objects directly cast to numeric values.
}
\examples{

# base R object
a <- 1:10
names(a) <- seq(10, 100, 10)
nums(a)

# XArray
data(exemplar)
colnums(exemplar)
rownums(exemplar)
}
