% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleocoastlines}
\alias{paleocoastlines}
\title{PaleoMAP PaleoCoastlines (excerpt)}
\format{
A \code{\link[via:SfArray-class]{SfArray}} with 3 continental margin and 3 paleocoastline layers (3 rows and 2 columns).
}
\source{
Kocsis, A. T., & Scotese, C. R. (2020). PaleoMAP PaleoCoastlines data [Data set]. Zenodo. https://doi.org/10.5281/zenodo.3903164
}
\usage{
data(paleocoastlines)
}
\description{
A dataset containing the coastline reconstructions based on the PaleoMAP PaleoDEMS \url{https://www.earthbyte.org/paleodem-resource-scotese-and-wright-2018/} and the Paleobiology Database \url{https://paleobiodb.org} for 0, 10 and 20Ma.
}
\details{
This is version v7. The article describing the entire set is under review. Once that is published, the entire dataset will be available.
}
\keyword{datasets}
