% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-rast.R
\docType{data}
\name{project}
\alias{project}
\alias{project,RasterArray-method}
\title{Project a '\code{\link[via:RasterArray-class]{RasterArray}}'-class object}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
project

\S4method{project}{RasterArray}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[via:RasterArray-class]{RasterArray}} object to project.}

\item{y}{A \code{\link[via:RasterArray-class]{RasterArray}} the same options as in \code{\link[terra]{project}}.}

\item{...}{additional arguments as for \code{\link[terra]{project}}.}
}
\value{
A projected \code{\link[via:RasterArray-class]{RasterArray}}-class object.
}
\description{
The method implemets the \code{\link[terra]{project}} function for '\code{\link[via:RasterArray-class]{RasterArray}}'-class objects.

The method is inherited from the '\code{\link[terra:rast]{SpatRaster}}' class. See \code{\link[terra:project]{project}} for details.
}
\examples{
# project first three to mollweide
ex <- rastex() 
mollEx <- project(ex[1:3], y="ESRI:54009")
}
\keyword{datasets}
