% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-base.R
\name{proxy}
\alias{proxy}
\alias{proxy,VirtualArray-method}
\title{The proxy of an from a class derived from '\code{\link[via:XArray-class]{VirtualArray}}'}
\usage{
proxy(x, ...)

\S4method{proxy}{VirtualArray}(x)
}
\arguments{
\item{x}{\code{\link[via:XArray-class]{XArray}}, \code{\link[via:RasterArray-class]{RasterArray}} or \code{\link[via:SfArray-class]{SfArray}} object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{vector}, \code{matrix} or \code{array} of characters representing the \code{\link[via:XArray-class]{VirtualArray}} structure.
}
\description{
This function returns an object that symbolizes the structure of layers in the '\code{\link[via:XArray-class]{XArray}}', '\code{\link[via:RasterArray-class]{RasterArray}}' or '\code{\link[via:SfArray-class]{SfArray}}'.
}
\details{
The \code{proxy} method wraps the names of layers in the \code{@stack} using the \code{@index} slot of the '\code{\link[via:XArray-class]{VirtualArray}}'.
}
\examples{
data(exemplar)
proxy(exemplar)

data(paleocoastlines)
proxy(paleocoastlines)
}
