% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rastex}
\alias{rastex}
\title{Procedural example structures to demonstrate the capabilities of the RasterArray-class}
\usage{
rastex()
}
\description{
Procedural example structures to demonstrate the capabilities of the RasterArray-class
}
\examples{
# create example
example <- rastex()

# subset - single bracket
example['b']

# subset - single bracket
example[c(4, 6)]

# subset - double bracket
example[["layer_2"]]
}
