% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-rast.R
\name{xres}
\alias{xres}
\alias{yres}
\alias{res}
\alias{xres,RasterArray-method}
\alias{yres,RasterArray-method}
\alias{res,RasterArray-method}
\title{Resolution of a '\code{\link[via:RasterArray-class]{RasterArray}}'-class object}
\usage{
\S4method{xres}{RasterArray}(x)

\S4method{yres}{RasterArray}(x)

\S4method{res}{RasterArray}(x)
}
\arguments{
\item{x}{a \code{RasterArray}-class object.}
}
\value{
A \code{numeric} vector.
}
\description{
The methods are inherited from the '\code{\link[terra:rast]{SpatRaster}}' class, see \code{\link[terra]{res}}. Replacement is not allowed.
}
\examples{
ex <- rastex() 
res(ex)
yres(ex)
xres(ex)
}
