% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-rast.R
\name{resample}
\alias{resample}
\alias{resample,RasterArray,ANY-method}
\alias{resample,RasterArray-method}
\title{Resampling a '\code{\link[via:RasterArray-class]{RasterArray}}'-class object}
\usage{
\S4method{resample}{RasterArray,ANY}(x, y, ...)
}
\arguments{
\item{x}{a \code{\link[via:RasterArray-class]{RasterArray}}-class object.}

\item{y}{The y argument of the \code{\link[terra:resample]{resample}} function.}

\item{...}{arguments passed to the \code{\link[terra:resample]{resample}} function.}
}
\value{
A resampled \code{\link[via:RasterArray-class]{RasterArray}}-class object.
}
\description{
The method is inherited from the '\code{\link[terra:rast]{SpatRaster}}' class.
}
\examples{
ex <- rastex()
if(requireNamespace("terra", quietly=TRUE)){
  template <- terra::rast(res=5)
  resampled <- resample(ex, template)
}
}
