% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterArray-rast.R
\name{rotate}
\alias{rotate}
\alias{rotate,RasterArray-method}
\title{Rotate a '\code{\link[via:RasterArray-class]{RasterArray}}'-class object}
\usage{
\S4method{rotate}{RasterArray}(x, ...)
}
\arguments{
\item{x}{(\code{\link[via:RasterArray-class]{RasterArray}}) Object.}

\item{...}{Additional arguments passed to the \code{\link[terra]{rotate}} function.}
}
\value{
A \code{\link[via:RasterArray-class]{RasterArray}}-class object.
}
\description{
The method is inherited from the '\code{\link[terra:rast]{SpatRaster}}' class.
}
