% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SfArray-sf.R
\name{st_crs}
\alias{st_crs}
\alias{st_crs.SfArray}
\title{Coordinate reference system of an '\code{\link[via:SfArray-class]{SfArray}}'-class object}
\usage{
\method{st_crs}{SfArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[sf:sf]{sf}}-class object.}

\item{...}{arguments passed to the \code{\link[sf:st_crs]{st_crs}} function.}
}
\value{
An aggregated \code{\link[via:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the '\code{\link[sf:sf]{sf}}' class.
}
\examples{
data(paleocoastlines)
crs <- st_crs(paleocoastlines)
}
