% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/SfArray-sf.R
\name{st_transform}
\alias{st_transform}
\alias{st_transform.SfArray}
\title{Projection change a \code{\link[genarray:SfArray-class]{SfArray}} object}
\usage{
\method{st_transform}{SfArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[sf:sf-class]{sf}}-class object.}

\item{...}{arguments passed to the \code{\link[raster]{extent}} function.}
}
\value{
An \code{\link[chronosphere:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the \code{\link[sf:sf]{sf}} class.
}
\examples{
data(dems)
agg <- ext(dems)
}
