% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-base.R
\name{t,VirtualArray-method}
\alias{t,VirtualArray-method}
\title{Transpose a \code{\link{RasterArray}} or \code{\link{SpatialArray}} object}
\usage{
\S4method{t}{VirtualArray}(x)
}
\arguments{
\item{x}{A \code{\link{RasterArray}} or \code{\link{SpatialArray}} class object.}
}
\value{
A \code{\link{RasterArray}} or \code{\link{SpatialArray}} class object.
}
\description{
Transpose a \code{\link{RasterArray}} or \code{\link{SpatialArray}} object
}
\examples{
data(dems)
t(dems)
data(coasts)
t(coasts)
}
