% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{ncol,VirtualArray-method}
\alias{ncol,VirtualArray-method}
\alias{nrow,VirtualArray-method}
\title{Number of columns and rows of a \code{\link[via:XArray-class]{VirtualArray}}-derived class object.}
\usage{
\S4method{ncol}{VirtualArray}(x)

\S4method{nrow}{VirtualArray}(x)
}
\arguments{
\item{x}{A \code{\link[via:XArray-class]{VirtualArray}}-derived class object.}
}
\value{
A \code{numeric} value of the number of columns and rows.
}
\description{
Unlike the \code{ncol} and \code{nrow} functions of the terra package, this function returns the number of columns and rows of the \code{VirtualArray}-derived container, rather than the dimensions of the contained \code{SpatRaster}s.
}
\examples{
data(paleocoastlines)
ncol(paleocoastlines)
nrow(paleocoastlines)
}
