% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{colnames,VirtualArray-method}
\alias{colnames,VirtualArray-method}
\alias{colnames<-,VirtualArray-method}
\title{Column names of two-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class object.}
\usage{
\S4method{colnames}{VirtualArray}(x)

\S4method{colnames}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link[via:XArray-class]{VirtualArray}}-derived class object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of column names or \code{NULL}.
}
\description{
Get or set the column names of two-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class objects
}
\examples{
data(paleocoastlines)
colnames(paleocoastlines)
colnames(paleocoastlines) <- c("a", "b")
}
