% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-combine.R
\name{c,XArray-method}
\alias{c,XArray-method}
\title{Combine GenArray and one-dimensional RasterArrays and Spatial* objects with one-dimensiooal SpatialArrays}
\usage{
\S4method{c}{XArray}(x, ...)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterArray} objects or Spatial* and \code{\link{SpatialArray}} objects to combine.}

\item{...}{additional objects to combine.}
}
\value{
A \code{\link{RasterArray}} or \code{\link{SpatialArray}}class object.
}
\description{
Methods sequences that start with an NA do not yet work.
}
