% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{dim,VirtualArray-method}
\alias{dim,VirtualArray-method}
\title{Dimensions of \code{\link[via:XArray-class]{VirtualArray}}-derived class objects}
\usage{
\S4method{dim}{VirtualArray}(x)
}
\arguments{
\item{x}{A \code{\link[via:XArray-class]{VirtualArray}}-derived class object.}
}
\value{
A \code{numeric} vector.
}
\description{
The function returns the dimensions of the array in which elements are organized.
}
\examples{
data(exemplar)
dim(exemplar)
}
