% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-xattrib.R
\name{dimlayer}
\alias{dimlayer}
\alias{dimlayer,RasterArray-method}
\title{Dimensions of layers in a '\code{\link[via:XArray-class]{VirtualArray}}'-class object}
\usage{
dimlayer(x, ...)

\S4method{dimlayer}{RasterArray}(x)
}
\arguments{
\item{x}{A \code{\link[via:XArray-class]{VirtualArray}} class object.}

\item{...}{additional arguments passed to class-specific methods.}
}
\value{
A \code{numeric} vector with the number of rows and columns in the \code{\link[via:XArray-class]{VirtualArray}}s.
}
\description{
The funcion will return the dimensions '\code{\link[terra:rast]{SpatRaster}}'-class layers.
}
