% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{dimnames,VirtualArray-method}
\alias{dimnames,VirtualArray-method}
\alias{dimnames<-,VirtualArray-method}
\title{Names of a multidimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class object.}
\usage{
\S4method{dimnames}{VirtualArray}(x)

\S4method{dimnames}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link{RasterArray}} or \code{\link[via:SfArray-class]{SfArray}} object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{list} of \code{character} vectors or \code{NULL}.
}
\description{
Get or set the dimnames of multidimensional \code{\link[via:XArray-class]{VirtualArray}}-derived class object.
}
\examples{
ex <- rastex()
dimnames(ex)
data(paleocoastlines)
dimnames(paleocoastlines)
dimnames(paleocoastlines)[[2]] <- c("first", "second")
names(dimnames(paleocoastlines)) <- c("age", "type")
}
