% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XArray-replace.R
\name{[[<-,XArray-method}
\alias{[[<-,XArray-method}
\title{Replace \code{RasterLayer}s in a \code{\link{RasterArray}} object and \code{Spatial*} objects in a \code{\link{SpatialArray}} object.}
\usage{
\S4method{[[}{XArray}(x, i) <- value
}
\arguments{
\item{x}{\code{\link{RasterArray}} or \code{\link{SpatialArray}} object.}

\item{i}{subscript of layers to replace.}

\item{value}{\code{character} vector.}
}
\value{
None.
}
\description{
Double bracket \code{'[['} refers to layers' name in the \code{RasterStack} of the \code{RasterArray} and the \code{\link{SpatialStack}} of the \code{\link{SpatialArray}}. Use single brackets to replace elements based on their position in the \code{RasterArray}/\code{\link{SpatialArray}}.
}
