% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exemplar}
\alias{exemplar}
\title{Example XArray object}
\format{
: \code{\link[via:XArray-class]{XArray}} with 3 sample sizes (rows), and 4 different seeds (column).
}
\usage{
data(exemplar)
}
\description{
An XArray of \code{data.frames} made from a single \code{data.frame} with random sampling. The original object had two columns, the first (\code{x}) an integer seqence \code{1:100}, the second \code{y} a variable produced with \code{0.5 * x -30 + N(0,10)}.
}
\keyword{datasets}
