% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/RasterArray-rast.R
\name{ext}
\alias{ext}
\alias{ext,RasterArray-method}
\title{Extent of a \code{\link[chronosphere:RasterArray-class]{RasterArray}} object}
\usage{
\S4method{ext}{RasterArray}(x, ...)
}
\arguments{
\item{x}{a \code{\link[chronosphere:RasterArray-class]{RasterArray}}-class object.}

\item{...}{arguments passed to the \code{\link[raster]{extent}} function.}
}
\value{
An aggregated \code{\link[chronosphere:RasterArray-class]{RasterArray}} class object.
}
\description{
The method is inherited from the \code{\link[raster:raster]{RasterStack}} class.
}
\examples{
data(dems)
agg <- ext(dems)
}
