% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rastex}
\alias{rastex}
\title{Procedural example structure to demonstrate the capabilities of the '\code{\link[via:RasterArray-class]{RasterArray}}' class}
\usage{
rastex()
}
\value{
A two-dimensional \code{\link[via:RasterArray-class]{RasterArray}}-class object, with three rows and four columns.
}
\description{
Binary versions of \code{\link[terra:rast]{SpatRaster}}-class objects are problematic, this function is used to instantiate a \code{\link[via:RasterArray-class]{RasterArray}} example.
}
\examples{
# create example
example <- rastex()

# subset - single bracket
example['b']

# subset - single bracket
example[c(4, 6)]

# subset - double bracket
example[["layer_2"]]
}
