% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-xattrib.R
\name{rownames,VirtualArray-method}
\alias{rownames,VirtualArray-method}
\alias{rownames<-,VirtualArray-method}
\title{Row names of two-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class objects.}
\usage{
\S4method{rownames}{VirtualArray}(x)

\S4method{rownames}{VirtualArray}(x) <- value
}
\arguments{
\item{x}{\code{\link[via:XArray-class]{VirtualArray}}-class object.}

\item{value}{\code{character} vector.}
}
\value{
A \code{character} vector of row names or \code{NULL}.
}
\description{
Get or set the row names of two-dimensional '\code{\link[via:XArray-class]{VirtualArray}}'-derived class object
}
\examples{
data(paleocoastlines)
rownames(paleocoastlines)
rownames(paleocoastlines) <- paste(rownames(paleocoastlines), "Ma")
}
