% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirtualArray-base.R
\name{t,VirtualArray-method}
\alias{t,VirtualArray-method}
\title{Transpose a \code{\link[via:XArray-class]{VirtualArray}} object}
\usage{
\S4method{t}{VirtualArray}(x)
}
\arguments{
\item{x}{A \code{\link[via:XArray-class]{VirtualArray}}-class object.}
}
\value{
A \code{\link[via:XArray-class]{VirtualArray}}-class object.
}
\description{
Transpose a \code{\link[via:XArray-class]{VirtualArray}} object
}
\examples{
data(exemplar)
t(exemplar)
data(paleocoastlines)
t(paleocoastlines)
}
